Added proper short typecasting to compile properly with javac

+++ db/com/cdegroot/db/recman/FreeLogicalRowIdPage.java
@@ -1,5 +1,5 @@
 /*
- *  $Id: FreeLogicalRowIdPage.java,v 1.1 2000/04/03 12:13:48 cdegroot Exp $
+ *  $Id: FreeLogicalRowIdPage.java,v 1.2 2000/04/11 06:07:42 boisvert Exp $
  *
  *  Logical row id pages
  *
@@ -7,8 +7,8 @@
  *  Copyright (C) 1999,2000 Cees de Groot <cg@cdegroot.com>
  *
  *  This library is free software; you can redistribute it and/or
- *  modify it under the terms of the GNU Library General Public License 
- *  as published by the Free Software Foundation; either version 2 
+ *  modify it under the terms of the GNU Library General Public License
+ *  as published by the Free Software Foundation; either version 2
  *  of the License, or (at your option) any later version.
  *
  *  This library is distributed in the hope that it will be useful,
@@ -16,8 +16,8 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *  Library General Public License for more details.
  *
- *  You should have received a copy of the GNU Library General Public License 
- *  along with this library; if not, write to the Free Software Foundation, 
+ *  You should have received a copy of the GNU Library General Public License
+ *  along with this library; if not, write to the Free Software Foundation,
  *  Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  */
 package com.cdegroot.db.recman;
@@ -31,10 +31,10 @@ package com.cdegroot.db.recman;
 class FreeLogicalRowIdPage extends PageHeader {
     // offsets
     private static final short O_COUNT = PageHeader.SIZE; // short count
-    static final short O_FREE = O_COUNT + Magic.SZ_SHORT;
-    static final short ELEMS_PER_PAGE = 
-	(RecordFile.BLOCK_SIZE - O_FREE) / PhysicalRowId.SIZE;
-    
+    static final short O_FREE = (short)(O_COUNT + Magic.SZ_SHORT);
+    static final short ELEMS_PER_PAGE = (short)
+	((RecordFile.BLOCK_SIZE - O_FREE) / PhysicalRowId.SIZE);
+
     // slots we returned.
     final PhysicalRowId[] slots = new PhysicalRowId[ELEMS_PER_PAGE];
 
@@ -90,11 +90,11 @@ class FreeLogicalRowIdPage extends PageHeader {
     boolean isFree(int slot) {
 	return !isAllocated(slot);
     }
-    
-    
+
+
     /** Returns the value of the indicated slot */
     PhysicalRowId get(int slot) {
-	if (slots[slot] == null) 
+	if (slots[slot] == null)
 	    slots[slot] = new PhysicalRowId(block, slotToOffset(slot));;
 	return slots[slot];
     }
@@ -104,8 +104,8 @@ class FreeLogicalRowIdPage extends PageHeader {
 	return (short) (O_FREE +
 	    (slot * PhysicalRowId.SIZE));
     }
-    
-    /** 
+
+    /**
      *  Returns first free slot, -1 if no slots are available
      */
     int getFirstFree() {
